Positions assigned to nonexistent objects
...because that's how it's supposed to work apparently!

The original spec says: "xpos,ypos: Position of object (from object mark 0). Max
256 positions are described on KISS/GS2. (- >7 KISS/GS Ranks) '*' means a
non-existent object." I misunderstood this to mean that the objects are ordered
from zero and '*' is assigned to objects that don't *appear* in that set. But
really, it's just what the spec says. The positions start with the position of
object 0, and positions are assigned for every integer until there are no more
objects. Weird, but every set I've tried (that rendered at all) has good
positions now!