Port cel2pnm to Haskell
Fixes #41 (Port cel2pnm to Haskell).

This commit provides a Haskell implementation of cel2pnm. In addition,
it removes the need for an external program, pnmtopng, by using the
JuicyPixels package to directly convert cels to PNGs. Thus, the entire
cel to PNG pipeline is now in pure Haskell. There is no longer a need
for a separate C program (cel2pnm) that has to be manually compiled and
placed in the PATH enviroment variable, easing the installation process
for new contributors.

However, the current Haskell implementation suffers from poor
performance, taking as much as three times longer to convert cels than
the previous combination of cel2pnm and pnmtopng. This will hopefully be
resolved in future commits.

Support for palette groups and Cherry KiSS were omitted in order to keep
the scope of the issue manageable. They will likely be added in the
future.