Fix calculation of pixel data size for 4bpp cels
The parsing function was incorrectly told to parse twice as many bytes
as needed. Since there aren't actually that many bytes in the cel file,
it resulted in an error when the function tried to request non-existent
bytes for parsing.

The miscalculation stemmed from forgetting that unlike the 8bpp format,
there're actually two pixels/colors stored in a single byte. Using the
same calculation for both 8bpp and 4bpp cels led to parsing twice as
many bytes in the latter case. We fix this by splitting the calculations
for 8bpp and 4bpp cels, making sure that the right formula is used for
each type of cel.

In addition, the buggy code failed to take into account the padding
pixels when a 4bpp cel's width is odd, which is necessary to ensure that
the correct pixel data size is computed. This is now handled properly.