Improve cel parsing performance
It seems that the cause of the slow performance can be attributed to a
space leak arising from the excessive laziness of list node allocations.

In the original code, the parser was returning a list of lists of one or
two elements, each representing an index into the palette. This was done
because the parser needed to return two bytes when parsing a 4bpp cel.
The resulting list would look something like [[0xAB, 0xCD], [0x10,
0x20]] for 4bpp cels, and [[0xAB], [0xCD], [0x10], [0x20]] for 8bpp
cels. Then, it would be concatenated before being returned to the
calling function.

This meant that much of the time and memory was spent on creating the
list, and then garbage collecting it. We can see this in [1], where GC
takes up 46.3% of the total running time when converting Arwen's
base.cel.

Our solution removes the use of BinaryParser from the parse4bpp and
parse8bpp functions, and instead parses the data manually. In the 8bpp
case, we need only return the bytes unmodified. For 4bpp cels, we use a
ByteString builder to accumulate the parsed bytes because it's more
efficient to build the final return value from a sequence of parsed
bytes this way (concatenation is an O(1) operation).

The improvement from this change is fairly noticable. Converting all 167
cels in the Arwen set to PNGs now takes about 1.1 times as long as
cel2pnm + pnmtopng, as measured by the "bench" benchmarking program
(https://hackage.haskell.org/package/bench). Previously, it took almost
3 times as long. Memory usage also improved significantly -- it now runs
in a constant 2 MB of space when converting Arwen's base.cel, compared
to 52 MB previously (see [1] and [2]).

When converting the Angel Lily set (mvks02st) which contains 92 4bpp
cels, it also took about 1.1 times as long. While the 4bpp parsing code
is currently broken, the fixed code should still perform similarly.
However, more testing is needed to see if further changes are required
to keep the performance competitive with cel2pnm + pnmtopng.

[1]: Runtime statistics for the original code when converting base.cel
from the Arwen set.

     208,448,728 bytes allocated in the heap
     101,407,216 bytes copied during GC
      25,268,880 bytes maximum residency (7 sample(s))
         122,224 bytes maximum slop
              52 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0       186 colls,     0 par    0.060s   0.075s     0.0004s    0.0158s
  Gen  1         7 colls,     0 par    0.003s   0.004s     0.0006s    0.0012s

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time    0.072s  (  0.076s elapsed)
  GC      time    0.063s  (  0.080s elapsed)
  RP      time    0.000s  (  0.000s elapsed)
  PROF    time    0.000s  (  0.000s elapsed)
  EXIT    time    0.000s  (  0.000s elapsed)
  Total   time    0.135s  (  0.156s elapsed)

  %GC     time      46.3%  (51.0% elapsed)

  Alloc rate    2,890,143,752 bytes per MUT second

  Productivity  53.5% of total user, 48.8% of total elapsed

[2]: Runtime statistics for the new code when converting base.cel from
the Arwen set.

     118,317,816 bytes allocated in the heap
         380,296 bytes copied during GC
         495,704 bytes maximum residency (2 sample(s))
          36,776 bytes maximum slop
               2 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0       113 colls,     0 par    0.001s   0.001s     0.0000s    0.0003s
  Gen  1         2 colls,     0 par    0.000s   0.000s     0.0001s    0.0002s

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time    0.046s  (  0.050s elapsed)
  GC      time    0.001s  (  0.001s elapsed)
  RP      time    0.000s  (  0.000s elapsed)
  PROF    time    0.000s  (  0.000s elapsed)
  EXIT    time    0.000s  (  0.000s elapsed)
  Total   time    0.047s  (  0.052s elapsed)

  %GC     time       1.2%  (2.1% elapsed)

  Alloc rate    2,586,860,291 bytes per MUT second

  Productivity  98.4% of total user, 97.6% of total elapsed