Fix color conversion (12-bit to 24-bit)
The current code did not properly map the 12 bits per pixel (bpp) colors
to the 24 bpp color space. It used the raw 4-bit values (0x00 to 0xFF)
for each color component (red, green, and blue) to construct the final
24 bpp color. The 24 bpp colors created thus ranged from rgb(0, 0, 0),
rgb(1, 1, 1), rgb(2, 2, 2), and so on, up to rgb(15, 15, 15). In other
words, the code created images that were black.

To fix this, we need to multiply each color component with a constant
number in order to obtain the correct 24 bpp color. We use 17 as the
multiplier, which GnomeKiSS also uses. The resulting 24 bpp colors now
range from rgb(0, 0, 0), rgb(17, 17, 17), rgb(34, 34, 34), ..., to
rgb(255, 255, 255), correcting the color deficiency in the buggy code.

Note that the choice of 17 as the multiplier differs from cel2pnm's.
cel2pnm uses 16, and the effect that this has is to make the colors a
slightly darker shade (approximately a 6% difference: 100 - 16 / 17 *
100 = 5.88).