Parse all palette groups
The current implementation returns a parse error if there are trailing
bytes in a KCF file. This restriction leads to a failure when parsing
valid, well-formed KCF files that have multiple palette groups (notably
old-style KCF files) because we only expect and parse a single group.

The solution adopted in this patch is to parse all palette groups in
both old- and new- style KCF files, but return only the first group.
This was done because complete support for palette groups has yet to be
implemented.

The tests were also updated to reflect the changes in the code.