Convert cels to palette-based PNGs
As pointed out by @emhoracek in #41, the cumulative difference in sizes
between paletted PNGs produced by pnmtopng and the RGBA truecolor ones
created by the Haskell port of cel2pnm "could feel very significant for
large sets," even more so when using Smooch "on a mobile phone plan."

Thus, we change the implementation to convert cels to palette-based PNGs
with transparency instead of RGBA truecolor. When the current code was
written, JuicyPixels had no support for RGBA palettes, so we did not go
that route then. This commit therefore requires JuicyPixels >= 3.2.9.1,
which is the first version that has the feature we need.

Although this patch reduces the sizes of the PNGs, it still does not
approach that of pnmtopng's. The disparity appears to be caused by
different bit depths. Taking washu.cel from the washua set as an
example, pnmtopng generates a PNG that has a bit depth of 4 and 11
palette entries, whereas the Haskell version has a bit depth of 8 and
256 palette entries. It would seem that pnmtopng only adds the colors
actually used by a cel to the PNG's palette, allowing it to use a
smaller bit depth and ultimately resulting in a smaller PNG.

This is a potential area for further improvements.

For reference, the sizes of washu.png are as follows:

    pnmtopng: 3.42 KB
    old Haskell code: 5.6 KB
    new Haskell code: 4.46 KB