Allow parsing of valid KCF files with trailing bytes
There are sets like "Momoko-hime" (ks_mmkh) and "Magical Princess
Holy-Up" (chakiss) with valid KCF files that do not strictly conform to
the KiSS General Specification. Instead of the expected file size of 320
bytes for old-style KCF files, the aforementioned sets have 512-byte
ones.

Apart from the extra 192 bytes, there doesn't seem to be anything wrong
with them. Loosen the restriction on trailing bytes in KCF files to
avoid returning an error with otherwise valid sets.

The tests were also updated to reflect the changes in the code.