Replace Data.List partial functions with safe variants
Use wrappers from the Safe library to avoid crashing the application at
runtime if partial functions are given arguments that break their
invariants.

Default values for the wrappers are provided to represent a "best
effort" at converting a cel to PNG, even if it results in a corrupted
image. The rationale for this is to allow the user to at least see
something on the screen instead of an error message, especially if only
a small minority of cels are problematic (using EitherT to handle errors
would abort the entire conversion process, even for good cels).