Redirect new users to their user page
We simplify things for new users by automatically logging them in and
redirecting them to their user page after account creation. This saves
them the step of having to return to the home page to sign in.

This closes #69 (Log user in and redirect to user page on account
creation).